Refactored contexts into hooks.
