+++ src/components/Menu.tsx
@@ -1,13 +1,8 @@
-import { type FC } from 'react';
+import { useMemo, type FC } from 'react';
 import { a, config, useSpring } from '@react-spring/web';
 import { menuBox } from 'styles/legacy';
 import { MenuItem } from 'components/MenuItem';
-import helloIcon from 'assets/icons/hello.svg';
-import storyIcon from 'assets/icons/bag.svg';
-import brainIcon from 'assets/icons/brain.svg';
-import livesIcon from 'assets/icons/lives.svg';
-import learnIcon from 'assets/icons/dots.svg';
-import { useMenuSnapshot } from 'context/MenuContext';
+import { items } from '../data/MenuData';
 export const Menu: FC = () => {
   const [{ width }] = useSpring(
     {
@@ -17,13 +12,6 @@ export const Menu: FC = () => {
     },
     [],
   );
-  const items = [
-    { to: () => useMenuSnapshot().hello.absoluteTop, icon: helloIcon, alt: 'Hello' },
-    { to: () => useMenuSnapshot().story.absoluteTop, icon: storyIcon, alt: 'History' },
-    { to: () => useMenuSnapshot().brain.absoluteTop, icon: brainIcon, alt: 'Skills' },
-    { to: () => useMenuSnapshot().lives.absoluteTop, icon: livesIcon, alt: 'Location' },
-    { to: () => useMenuSnapshot().learn.absoluteTop, icon: learnIcon, alt: 'Learn More' },
-  ];
   return (
     <a.div className={menuBox} style={{ width }}>
       {items.map((props, index) => (
